+++ data/src/main/java/justbucket/videolib/data/model/FilterEntity.kt
@@ -1,6 +1,6 @@
 package justbucket.videolib.data.model
 
 data class FilterEntity(val text: String = "",
-                        val sources: List<Int> = emptyList(),
+                        val sources: List<Short> = emptyList(),
                         val allAnyCheck: Boolean = false,
                         val tags: List<TagEntity> = emptyList())
