+++ domain/src/main/kotlin/justbucket/videolib/domain/model/Filter.kt
@@ -1,6 +1,6 @@
 package justbucket.videolib.domain.model
 
 data class Filter(val text: String,
-                  val sources: List<Int>,
+                  val sources: List<Short>,
                   val allAnyCheck: Boolean,
                   val tags: List<Tag>)
