+++ domain/src/main/kotlin/justbucket/videolib/domain/model/Video.kt
@@ -4,5 +4,5 @@ data class Video(val id: Long,
                  val title: String,
                  val videoPath: String,
                  val thumbPath: String,
-                 val source: Int,
+                 val source: Short,
                  val tags: List<Tag>)
